+++ lua/plugins/lang/svelte.lua
